requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-admin-interface==0.28.6
django-colorfield==0.11.0
idna==3.6
Pillow==10.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-slugify==8.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
ua-parser==0.18.0A
urllib3==1.26.1
user-agents==2.2.0
whitenoise==6.6.0
djangorestframework
